+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - [#9306](https://github.com/blockscout/blockscout/pull/9306) - Improve marking of failed internal transactions
 - [#9305](https://github.com/blockscout/blockscout/pull/9305) - Add effective gas price calculation as fallback
 - [#9300](https://github.com/blockscout/blockscout/pull/9300) - Fix read contract bug
+- [#9226](https://github.com/blockscout/blockscout/pull/9226) - Split Indexer.Fetcher.TokenInstance.LegacySanitize
 
 ### Chore